+++ test/test_helpers/nibble_starter_helper.rb
@@ -9,7 +9,8 @@ module NibbleStarterHelper
       "locales" => [ { "code" => "en", "default" => true } ], "reserved_paths" => [ "/admin" ],
       "assets" => { "presets" => { "og" => { "w" => 1200, "h" => 630, "fit" => "crop" } } } },
       themes_path: Rails.root.join("test/nibble_themes"),
-      site_schema_path: Rails.root.join("test/nibble_themes/no_site_schema"))
+      site_schema_path: Rails.root.join("test/nibble_themes/no_site_schema"),
+      published_path: Pathname(Dir.mktmpdir("nibble-published")))
     Nibble.reset_schema!
     Nibble.boot!
     Nibble::PageCache.store = ActiveSupport::Cache::MemoryStore.new